Made end-to-end test.    For Login, calendar, and logout
Made a end-to-end test that test whether a users login. By filling content in the formfields, and click on login button. Calendar By testing whether the date change when clicking on  'vorige week', or  'volgende week' and if that are the correct dates. And made logout, when user click on  there profile, and press logout. That they get navigated back to landing page.